# server/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EaaS Nexus Backend API",
    "version": "1.0.0",
    "description": "Backend API contract for EaaS Nexus mobile application."
  },
  "servers": [
    { "url": "https://server-plum-six.vercel.app", "description": "Production" },
    { "url": "http://localhost:3001", "description": "Development" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": { "type": "string" },
          "data": { "type": "object", "additionalProperties": true }
        },
        "required": ["success"]
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/energy/realtime/{propertyId}": {
      "get": {
        "tags": ["Energy"],
        "summary": "Get realtime energy data",
        "parameters": [
          {
            "in": "path",
            "name": "propertyId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Realtime energy",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          }
        }
      }
    },
    "/energy/stats/{propertyId}": {
      "get": {
        "tags": ["Energy"],
        "summary": "Get energy statistics",
        "parameters": [
          {
            "in": "path",
            "name": "propertyId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "query",
            "name": "period",
            "schema": {
              "type": "string",
              "enum": ["day", "week", "month", "year"]
            }
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": { "type": "string", "format": "date-time" }
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": { "type": "string", "format": "date-time" }
          }
        ],
        "responses": {
          "200": {
            "description": "Energy stats",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          }
        }
      }
    },
    "/energy/stream/{propertyId}": {
      "get": {
        "tags": ["Energy"],
        "summary": "Get Supabase realtime subscription config",
        "description": "Returns channel/table/filter info for client-side Supabase realtime subscription.",
        "parameters": [
          {
            "in": "path",
            "name": "propertyId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Realtime subscription config",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          }
        }
      }
    },
...
    "/billing/current/{propertyId}": {
      "get": {
        "tags": ["Billing"],
        "summary": "Get current bill",
        "parameters": [
          {
            "in": "path",
            "name": "propertyId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Current bill",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          }
        }
      }
    },
    "/billing/history/{propertyId}": {
      "get": {
        "tags": ["Billing"],
        "summary": "Get billing history",
        "parameters": [
          {
            "in": "path",
            "name": "propertyId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer", "default": 12 }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": { "type": "integer", "default": 0 }
          }
        ],
        "responses": {
          "200": {
            "description": "Billing history",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          }
        }
      }
    },
    "/billing/download/{billId}": {
      "get": {
        "tags": ["Billing"],
        "summary": "Download bill PDF",
        "parameters": [
          {
            "in": "path",
            "name": "billId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "PDF file",
            "content": { "application/pdf": {} }
          }
        }
      }
    },
    "/billing/admin/all": {
      "get": {
        "tags": ["Billing"],
        "summary": "Admin list all bills",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer", "default": 50 }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": { "type": "integer", "default": 0 }
          }
        ],
        "responses": {
          "200": {
            "description": "Bills list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          }
        }
      }
    },
    "/billing/admin/{billId}/adjust": {
      "patch": {
        "tags": ["Billing"],
        "summary": "Admin adjust bill",
        "parameters": [
          {
            "in": "path",
            "name": "billId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "usageCharge": 1500,
                "subscriptionFee": 2500,
                "taxes": 300,
                "status": "pending",
                "note": "Manual correction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bill adjusted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          }
        }
      }
    },
...
  }
}